requirements.txt
flickrapi==2.4.0
matplotlib==3.2.2
Pillow==7.2.0
pytest==5.4.3
requests==2.24.0
textblob==0.15.3
tweepy==3.9.0